yrs::any

Enum Any

pub enum Any {
    Null,
    Undefined,
    Bool(bool),
    Number(f64),
    BigInt(i64),
    String(Arc<str>),
    Buffer(Arc<[u8]>),
    Array(Arc<[Any]>),
    Map(Arc<HashMap<String, Any>>),
}

Any is an enum with a potentially associated value that is used to represent JSON values and supports efficient encoding of those values.

Variants

Null

Undefined

Bool(bool)

Number(f64)

BigInt(i64)

String(Arc<str>)

Buffer(Arc<[u8]>)

Array(Arc<[Any]>)

Map(Arc<HashMap<String, Any>>)

Implementations

impl Any

pub fn cast<T>(self) -> Result<T, Self>
where T: TryFrom<Any, Error = Any>,

pub fn decode<R: Read>(decoder: &mut R) -> Result<Self, Error>

pub fn encode<W: Write>(&self, encoder: &mut W)

pub fn from_json(src: &str) -> Result<Self, Error>

pub fn to_json(&self, buf: &mut String)