yrs

Module branch

Structs

Branch

Branch describes a content of a complex Yrs data structures, such as arrays or maps.

BranchPtr

A wrapper around Branch cell, supplied with a bunch of convenience methods to operate on both map-like and array-like contents of a Branch.

Hook

A descriptor used to reference to shared collections by their unique logical identifiers, which can be either Root-level collections or shared collections Nested into each other. It can be resolved from any shared reference using SharedRef::hook.

Nested

A logical reference used to represent a shared collection nested within another one. Unlike Root-level types which cannot be deleted and exist eternally, Nested collections can be added (therefore don’t exist prior their instantiation) and deleted (so that any SharedRef values referencing them become unsafe and can point to objects that no longer exists!).

Root

A logical reference to a root-level shared collection. It can be shared across different documents to reference the same logical type.

Enums

BranchID

An unique logical identifier of a shared collection. Can be shared across document boundaries to reference to the same logical entity across different replicas of a document.