yrs::branch

Struct Branch

pub struct Branch {
    pub block_len: u32,
    pub content_len: u32,
    /* private fields */
}

Branch describes a content of a complex Yrs data structures, such as arrays or maps.

Fields
block_len: u32

A length of an indexed sequence component of a current branch node. Map component elements are computed on demand.

content_len: u32
Implementations

impl Branch

pub fn new(type_ref: TypeRef) -> Box<Self>

pub fn is_deleted(&self) -> bool

pub fn id(&self) -> BranchID

pub fn as_subdoc(&self) -> Option<Doc>

pub fn type_ref(&self) -> &TypeRef

Returns an identifier of an underlying complex data type (eg. is it an Array or a Map).

pub fn len(&self) -> u32

Returns a length of an indexed sequence component of a current branch node. Map component elements are computed on demand.

pub fn content_len(&self) -> u32

pub fn observe<F>(&mut self, f: F) -> Subscription
where F: Fn(&TransactionMut<'_>, &Event) + 'static,

pub fn observe_with<F>(&mut self, key: Origin, f: F)
where F: Fn(&TransactionMut<'_>, &Event) + 'static,

pub fn unobserve(&mut self, key: &Origin) -> bool

pub fn observe_deep<F>(&self, f: F) -> Subscription
where F: Fn(&TransactionMut<'_>, &Events<'_>) + 'static,

pub fn observe_deep_with<F>(&self, key: Origin, f: F)
where F: Fn(&TransactionMut<'_>, &Events<'_>) + 'static,