yrs::branch

Struct BranchPtr

pub struct BranchPtr(/* private fields */);

A wrapper around Branch cell, supplied with a bunch of convenience methods to operate on both map-like and array-like contents of a Branch.