yrs

Enum Assoc

#[repr(i8)]
pub enum Assoc {
    After = 0,
    Before = -1,
}

Association type used by StickyIndex. In general StickyIndex refers to a cursor space between two elements (eg. “ab.c” where “abc” is our string and . is the StickyIndex placement). However in a situation when another peer is updating a collection concurrently, a new set of elements may be inserted into that space, expanding it in the result. In such case Assoc tells us if the StickyIndex should stick to location before or after referenced index.

Variants

After = 0

The corresponding StickyIndex points to space after the referenced ID.

Before = -1

The corresponding StickyIndex points to space before the referenced ID.