yrs

Enum Out

pub enum Out {
    Any(Any),
    YText(TextRef),
    YArray(ArrayRef),
    YMap(MapRef),
    YXmlElement(XmlElementRef),
    YXmlFragment(XmlFragmentRef),
    YXmlText(XmlTextRef),
    YDoc(Doc),
    UndefinedRef(BranchPtr),
}

Value that can be returned by Yrs data types. This includes Any which is an extension representation of JSON, but also nested complex collaborative structures specific to Yrs.

Variants

Any(Any)

Any value that it treated as a single element in its entirety.

YText(TextRef)

Instance of a TextRef.

YArray(ArrayRef)

Instance of an ArrayRef.

YMap(MapRef)

Instance of a MapRef.

YXmlElement(XmlElementRef)

Instance of a XmlElementRef.

YXmlFragment(XmlFragmentRef)

Instance of a XmlFragmentRef.

YXmlText(XmlTextRef)

Instance of a XmlTextRef.

YDoc(Doc)

Subdocument.

UndefinedRef(BranchPtr)

Instance of a shared collection of undefined type. Usually happens when it refers to a root type that has not been defined locally. Can also refer to a [WeakRef] if “weak” feature flag was not set.

Implementations

impl Out

pub fn cast<T>(self) -> Result<T, Self>
where T: TryFrom<Self, Error = Self>,

Attempts to convert current Out value directly onto a different type, as along as it implements TryFrom trait. If conversion is not possible, the original value is returned.

pub fn to_string<T: ReadTxn>(self, txn: &T) -> String

Converts current value into stringified representation.

pub fn try_branch(&self) -> Option<&Branch>