yrs::observer

Struct Observer

pub struct Observer<F> { /* private fields */ }

Data structure used to handle publish/subscribe callbacks of specific type. Observers perform subscriber changes in thread-safe manner, using atomic hardware intrinsics.

Implementations

impl<F> Observer<F>
where F: 'static,

pub fn new() -> Self

Creates a new Observer with no active callbacks.

pub fn has_subscribers(&self) -> bool

pub fn clean(&self)

Cleanup already released subscriptions. Whenever a Subscription is dropped, the callback is released. However, the weak reference to callback may still be kept around until it becomes touched by operations such as Observer::subscribe or [Observer::callbacks].

This method allows to perform stale callback cleanup without waiting for callbacks to be visited.

pub fn unsubscribe(&self, id: &Origin) -> bool

pub fn trigger<E>(&self, each: E)
where E: FnMut(&F),

Returns a snapshot of callbacks subscribed to this observer at the moment when this method has been called. This snapshot can be iterated over to get access to individual callbacks and trigger them.

pub fn subscribe_with(&self, id: Origin, callback: F)

Subscribes a callback parameter to a current Observer. Returns a subscription object which - when dropped - will unsubscribe current callback. If the id was already present in the observer, current callback will be ignored.

impl<F> Observer<F>
where F: 'static,

pub fn subscribe(&self, callback: F) -> Subscription