yrs::observer

Type Alias Subscription

pub type Subscription = Arc<dyn Drop + 'static>;

Subscription handle returned by Observer::subscribe methods, which will unsubscribe corresponding callback when dropped.

If you need to send the Subscription handle to another thread, you may wish to enable the sync feature such that Subscription implements Send+Sync.

Aliased Type

pub struct Subscription { /* private fields */ }