yrs

Struct StateVector

pub struct StateVector(/* private fields */);

State vector is a compact representation of all known blocks inserted and integrated into a given document. This descriptor can be serialized and used to determine a difference between seen and unseen inserts of two replicas of the same document, potentially existing in different processes.

Another popular name for the concept represented by state vector is Version Vector.

Implementations

impl StateVector

pub fn is_empty(&self) -> bool

Checks if current state vector contains any data.

pub fn len(&self) -> usize

Returns a number of unique clients observed by a document, current state vector corresponds to.

pub fn new( map: HashMap<ClientID, u32, BuildHasherDefault<ClientHasher>>, ) -> Self

pub fn contains(&self, id: &ID) -> bool

Checks if current state vector includes given block identifier. Blocks, which identifiers can be found in a state vectors don’t need to be encoded as part of an update, because they were already observed by their remote peer, current state vector refers to.

pub fn contains_client(&self, client_id: &ClientID) -> bool

pub fn get(&self, client_id: &ClientID) -> u32

Get the latest clock sequence number value for a given client_id as observed from the perspective of a current state vector.

pub fn inc_by(&mut self, client: ClientID, delta: u32)

Updates a state vector observed clock sequence number for a given client by incrementing it by a given delta.

pub fn set_min(&mut self, client: ClientID, clock: u32)

Updates a state vector observed clock sequence number for a given client by setting it to a minimum value between an already present one and the provided clock. In case if state vector didn’t contain any value for that client, a clock value will be used.

pub fn set_max(&mut self, client: ClientID, clock: u32)

Updates a state vector observed clock sequence number for a given client by setting it to a maximum value between an already present one and the provided clock. In case if state vector didn’t contain any value for that client, a clock value will be used.

pub fn iter(&self) -> Iter<'_, ClientID, u32>

Returns an iterator which enables to traverse over all clients and their known clock values described by a current state vector.

pub fn merge(&mut self, other: Self)

Merges another state vector into a current one. Since vector’s clock values can only be incremented, whenever a conflict between two states happen (both state vectors have different clock values for the same client entry), a highest of these to is considered to be the most up-to-date.