yrs

Struct TransactionMut

pub struct TransactionMut<'doc> { /* private fields */ }

Read-write transaction. It can be used to modify an underlying state of the corresponding Doc. Read-write transactions require an exclusive access to document store - only one such transaction can be present per Doc at the same time (read-only Transactions are not allowed to coexists at the same time as well).

This transaction type stores the information about all of the changes performed in its scope. These will be used during TransactionMut::commit call to optimize metadata of incoming updates, triggering necessary event callbacks etc. For performance reasons it’s preferred to batch as many updates as possible using the same transaction.

In Yrs transactions are always auto-committing all of their changes when dropped. Rollbacks are not supported (if some operations needs to be undone, this can be achieved using [UndoManager])

Implementations

impl<'doc> TransactionMut<'doc>

pub fn doc(&self) -> &Doc

pub fn events(&self) -> Option<&StoreEvents>

pub fn events_mut(&mut self) -> &mut StoreEvents

pub fn before_state(&self) -> &StateVector

Corresponding document’s state vector at the moment when current transaction was created.

pub fn after_state(&self) -> &StateVector

State vector of the transaction after [Transaction::commit] has been called.

pub fn delete_set(&self) -> &DeleteSet

Data about deletions performed in the scope of current transaction.

pub fn origin(&self) -> Option<&Origin>

Returns origin of the transaction if any was defined. Read-write transactions can get an origin assigned via Transact::try_transact_mut_with/Transact::transact_mut_with methods.

pub fn changed_parent_types(&self) -> &[BranchPtr]

Returns a list of root level types changed in a scope of the current transaction. This list is not filled right away, but as a part of TransactionMut::commit process.

pub fn encode_update_v1(&self) -> Vec<u8>

Encodes changes made within the scope of the current transaction using lib0 v1 encoding.

Document updates are idempotent and commutative. Caveats:

It doesn’t matter in which order document updates are applied.

As long as all clients receive the same document updates, all clients end up with the same content.

Even if an update contains known information, the unknown information is extracted and integrated into the document structure.

pub fn encode_update_v2(&self) -> Vec<u8>

Encodes changes made within the scope of the current transaction using lib0 v2 encoding.

Document updates are idempotent and commutative. Caveats:

It doesn’t matter in which order document updates are applied.

As long as all clients receive the same document updates, all clients end up with the same content.

Even if an update contains known information, the unknown information is extracted and integrated into the document structure.

pub fn encode_update<E: Encoder>(&self, encoder: &mut E)

Encodes changes made within the scope of the current transaction.

Document updates are idempotent and commutative. Caveats:

It doesn’t matter in which order document updates are applied.

As long as all clients receive the same document updates, all clients end up with the same content.

Even if an update contains known information, the unknown information is extracted and integrated into the document structure.

pub fn apply_update(&mut self, update: Update) -> Result<(), UpdateError>

Applies a deserialized Update contents into a document owning current transaction. Update payload can be generated by methods such as TransactionMut::encode_diff or passed to Doc::observe_update_v1/Doc::observe_update_v2 callbacks. Updates are allowed to contain duplicate blocks (already presen in current document store) - these will be ignored.

Pending updates

Remote update integration requires that all to-be-integrated blocks must have their direct predecessors already in place. Out of order updates from the same peer will be stashed internally and their integration will be postponed until missing blocks arrive first.

pub fn commit(&mut self)

Commits current transaction. This step involves cleaning up and optimizing changes performed during lifetime of a transaction. Such changes include squashing delete sets data, squashing blocks that have been appended one after another to preserve memory and triggering events.

This step is performed automatically when a transaction is about to be dropped (its life scope comes to an end).

pub fn force_gc(&mut self)

Perform garbage collection of deleted blocks, even if a document was created with skip_gc option. This operation will scan over ALL deleted elements, NOT ONLY the ones that have been changed as part of this transaction scope.