yrs::sync

Module awareness

Structs

Awareness

The Awareness class implements a simple shared state protocol that can be used for non-persistent data like awareness information (cursor, username, status, ..). Each client can update its own local state and listen to state changes of remote clients.

AwarenessIter

AwarenessUpdate

A structure that represents an encodable state of an Awareness struct.

AwarenessUpdateEntry

A single client entry of an AwarenessUpdate. It consists of logical clock and JSON client state represented as a string.

AwarenessUpdateSummary

Summary of applying an AwarenessUpdate over Awareness::apply_update_summary method.

ClientState

Event

Event type emitted by an Awareness struct.

Enums

Error

Errors generated by an Awareness struct methods.