yrs::sync

Module protocol

Structs

DefaultProtocol

A default implementation of y-sync Protocol.

MessageReader

Since y-sync protocol enables for a multiple messages to be packed into a singe byte payload, MessageReader can be used over the decoder to read these messages one by one in iterable fashion.

Enums

Error

An error type returned in response from y-sync Protocol.

Message

SyncMessage

Constants

MSG_AUTH

Tag id for Message::Auth.

MSG_AWARENESS

Tag id for Message::Awareness.

MSG_QUERY_AWARENESS

Tag id for Message::AwarenessQuery.

MSG_SYNC

Tag id for Message::Sync.

MSG_SYNC_STEP_1

Tag id for SyncMessage::SyncStep1.

MSG_SYNC_STEP_2

Tag id for SyncMessage::SyncStep2.

MSG_SYNC_UPDATE

Tag id for SyncMessage::Update.

PERMISSION_DENIED

PERMISSION_GRANTED

Traits

AsyncProtocol

Trait implementing a y-sync protocol using awaitable transaction API. The default implementation can be found in DefaultProtocol, but its implementation steps can be potentially changed by the user if necessary.

Protocol

Trait implementing a y-sync protocol. The default implementation can be found in DefaultProtocol, but its implementation steps can be potentially changed by the user if necessary.