yrs

Type Alias Value

pub type Value = Out;

👎Deprecated since 0.19.0: Use yrs::Out instead

Aliased Type

pub enum Value {
    Any(Any),
    YText(TextRef),
    YArray(ArrayRef),
    YMap(MapRef),
    YXmlElement(XmlElementRef),
    YXmlFragment(XmlFragmentRef),
    YXmlText(XmlTextRef),
    YDoc(Doc),
    UndefinedRef(BranchPtr),
}

Variants

Any(Any)

Any value that it treated as a single element in its entirety.

YText(TextRef)

Instance of a TextRef.

YArray(ArrayRef)

Instance of an ArrayRef.

YMap(MapRef)

Instance of a MapRef.

YXmlElement(XmlElementRef)

Instance of a XmlElementRef.

YXmlFragment(XmlFragmentRef)

Instance of a XmlFragmentRef.

YXmlText(XmlTextRef)

Instance of a XmlTextRef.

YDoc(Doc)

Subdocument.

UndefinedRef(BranchPtr)

Instance of a shared collection of undefined type. Usually happens when it refers to a root type that has not been defined locally. Can also refer to a [WeakRef] if “weak” feature flag was not set.