yrs

Module types

Re-exports

pub use map::Map;

pub use map::MapRef;

pub use text::Text;

pub use text::TextRef;

Modules

array

map

text

xml

Structs

Events

EventsIter

Enums

Change

A single change done over an array-component of shared data type.

Delta

A single change done over a text-like types: Text or [XmlText].

EntryChange

A single change done over a map-component of shared data type.

Event

Generalized wrapper around events fired by specialized shared data types.

PathSegment

A single segment of a Path.

TypeRef

Constants

TYPE_REFS_ARRAY

Type ref identifier for an ArrayRef type.

TYPE_REFS_DOC

Type ref identifier for a [DocRef] type.

TYPE_REFS_MAP

Type ref identifier for a MapRef type.

TYPE_REFS_TEXT

Type ref identifier for a TextRef type.

TYPE_REFS_UNDEFINED

Placeholder type ref identifier for non-specialized AbstractType. Used only for root-level types which have been integrated from remote peers before they were defined locally.

TYPE_REFS_WEAK

Type ref identifier for a [WeakRef] type.

TYPE_REFS_XML_ELEMENT

Type ref identifier for a XmlElementRef type.

TYPE_REFS_XML_FRAGMENT

Type ref identifier for a XmlFragmentRef type. Used for compatibility.

TYPE_REFS_XML_HOOK

Type ref identifier for a [XmlHookRef] type. Used for compatibility.

TYPE_REFS_XML_TEXT

Type ref identifier for a XmlTextRef type.

Traits

AsPrelim

Trait which allows conversion back to a prelim type that can be used to create a new shared that’s a deep copy equivalent of a current type.

DeepObservable

Trait implemented by all Y-types, allowing for observing events which are emitted by nested types.

DefaultPrelim

Trait which allows to generate a Prelim-compatible type that - when integrated - will be converted into an instance of a current type.

GetString

Trait implemented by shared types to display their contents in string format.

Observable

RootRef

A subset of SharedRef used to mark collaborative collections that can be used as a root level collections. This includes common types like ArrayRef, MapRef, TextRef and XmlFragmentRef.

SharedRef

Common trait for shared collaborative collection types in Yrs.

ToJson

Type Aliases

Attrs

An alias for map of attributes used as formatting parameters by Text and [XmlText] types.

Path

A path describing nesting structure between shared collections containing each other. It’s a collection of segments which refer to either index (in case of Array or [XmlElement]) or string key (in case of Map) where successor shared collection can be found within subsequent parent types.