yrs::types::array

Struct ArrayRef

pub struct ArrayRef(/* private fields */);

A collection used to store data in an indexed sequence structure. This type is internally implemented as a double linked list, which may squash values inserted directly one after another into single list node upon transaction commit.

Reading a root-level type as an ArrayRef means treating its sequence components as a list, where every countable element becomes an individual entity:

JSON-like primitives (booleans, numbers, strings, JSON maps, arrays etc.) are counted individually.

Text chunks inserted by [Text] data structure: each character becomes an element of an array.

Embedded and binary values: they count as a single element even though they correspond of multiple bytes.

Like all Yrs shared data types, ArrayRef is resistant to the problem of interleaving (situation when elements inserted one after another may interleave with other peers concurrent inserts after merging all updates together). In case of Yrs conflict resolution is solved by using unique document id to determine correct and consistent ordering.

Example

use yrs::{Array, Doc, Map, MapPrelim, Transact, Any, any};
use yrs::types::ToJson;

let doc = Doc::new();
let array = doc.get_or_insert_array("array");
let mut txn = doc.transact_mut();

// insert single scalar value
array.insert(&mut txn, 0, "value");
array.remove_range(&mut txn, 0, 1);

assert_eq!(array.len(&txn), 0);

// insert multiple values at once
array.insert_range(&mut txn, 0, ["a", "b", "c"]);
assert_eq!(array.len(&txn), 3);

// get value
let value = array.get(&txn, 1);
assert_eq!(value, Some("b".into()));

// insert nested shared types
let map = array.insert(&mut txn, 1, MapPrelim::from([("key1", "value1")]));
map.insert(&mut txn, "key2", "value2");

assert_eq!(array.to_json(&txn), any!([
  "a",
  { "key1": "value1", "key2": "value2" },
  "b",
  "c"
]));