yrs::types

Trait AsPrelim

pub trait AsPrelim {
    type Prelim: Prelim<Return = Self>;

    // Required method
    fn as_prelim<T: ReadTxn>(&self, txn: &T) -> Self::Prelim;
}

Trait which allows conversion back to a prelim type that can be used to create a new shared that’s a deep copy equivalent of a current type.

Required Associated Types

type Prelim: Prelim<Return = Self>

Required Methods

fn as_prelim<T: ReadTxn>(&self, txn: &T) -> Self::Prelim

Converts current type contents into a Prelim type that can be used to create a new type that’s a deep copy equivalent of a current type.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.