yrs::types

Trait DefaultPrelim

pub trait DefaultPrelim {
    type Prelim: Prelim<Return = Self>;

    // Required method
    fn default_prelim() -> Self::Prelim;
}

Trait which allows to generate a Prelim-compatible type that - when integrated - will be converted into an instance of a current type.

Required Associated Types

type Prelim: Prelim<Return = Self>

Required Methods

fn default_prelim() -> Self::Prelim

Returns an instance of Prelim-compatible type, which will turn into reference of a current type after being integrated into the document store.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.