yrs::types

Trait GetString

pub trait GetString {
    // Required method
    fn get_string<T: ReadTxn>(&self, txn: &T) -> String;
}

Trait implemented by shared types to display their contents in string format.

Required Methods

fn get_string<T: ReadTxn>(&self, txn: &T) -> String

Displays the content of a current collection in string format.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.